Shield panel origin of url display and format url string test functions
